rename pch to ttnn-precompiled
